Refactored authorizedData object reconstruction for default steps
Previously authorizedData was built based on data from post from the pushed HTML object, which was susceptible to tampering without an additional layer of validations. Now only selected account and permission UUIDs are sent in the post and the selections are constructed in the backend based on necessary retrieved data passed along as consent meta data for persistence.